+++ src/groups/circle_group_sphere.jl
@@ -89,9 +89,17 @@ function _compose!(::_PlanarCircleGroup, k, g, h)
     return k
 end
 
-identity_element(::_PlanarCircleGroup) = [1.0, 0.0]
-function identity_element!(::_PlanarCircleGroup, p::Array{<:Any,1})
-    p = [1.0, 0.0]
+function identity_element(G::_PlanarCircleGroup)
+    return identity_element(G, AbstractVector)
+end
+function identity_element(::_PlanarCircleGroup, ::Type{<:AbstractVector})
+    return [1.0, 0.0]
+end
+function identity_element(::_PlanarCircleGroup, ::Type{<:AbstractVector{T}}) where {T}
+    return [one(T), zero(T)]
+end
+function identity_element!(::_PlanarCircleGroup, p::V) where {T,V<:AbstractVector{T}}
+    p .= [one(T), zero(T)]
     return p
 end