+++ src/groups/validation_group.jl
@@ -94,6 +94,10 @@ unwrap_validation(v) = v
 unwrap_validation(vTV::ValidationLieAlgebraTangentVector) = vTV.value
 unwrap_validation(vP::ValidationMPoint) = vP.value
 
+unwrap_validation_type(v) = v
+unwrap_validation_type(::Type{<:ValidationLieAlgebraTangentVector{T}}) where {T} = T
+unwrap_validation_type(::Type{<:ValidationMPoint{<:P}}) where {P} = P
+
 #
 #
 # An access helper function
@@ -377,7 +381,7 @@ function ManifoldsBase.hat(
     𝔤::LieAlgebra{𝔽,O,<:ValidationLieGroup}, c, T::Type; kwargs...
 ) where {𝔽,O<:AbstractGroupOperation}
     G = base_lie_group(𝔤).lie_group
-    X = hat(LieAlgebra(G), c, T)
+    X = hat(LieAlgebra(G), c, unwrap_validation_type(T))
     is_point(𝔤, X; widthin=hat, context=(:Output,), kwargs...)
     return ValidationLieAlgebraTangentVector(X)
 end
@@ -390,12 +394,11 @@ function ManifoldsBase.hat!(
     return X
 end
 
-function identity_element(G::ValidationLieGroup; kwargs...)
-    g = identity_element(G.lie_group)
+function identity_element(G::ValidationLieGroup, ::Type{T}; kwargs...) where {T}
+    g = identity_element(G.lie_group, unwrap_validation_type(T))
     is_point(G, g; widthin=identity_element, context=(:Output,), kwargs...)
     return ValidationMPoint(g)
 end
-
 function identity_element!(G::ValidationLieGroup, g; kwargs...)
     identity_element!(G.lie_group, unwrap_validation(g))
     is_point(G, g; widthin=identity_element, context=(:Output,), kwargs...)
@@ -627,13 +630,22 @@ function Base.log(
     return ValidationLieAlgebraTangentVector(X)
 end
 function Base.log(
-    G::ValidationLieGroup{𝔽,O}, e::Identity{O}, T::Type; kwargs...
-) where {𝔽,O<:AbstractGroupOperation}
+    G::ValidationLieGroup{𝔽,O}, e::Identity{O}, ::Type{T}; kwargs...
+) where {𝔽,O<:AbstractGroupOperation,T}
     X = log(G.lie_group, e, T)
     is_point(LieAlgebra(G), X; widthin=log, context=(:Output,), kwargs...)
     return ValidationLieAlgebraTangentVector(X)
 end
-
+function Base.log(
+    G::ValidationLieGroup{𝔽,O},
+    e::Identity{O},
+    ::Type{<:ValidationLieAlgebraTangentVector{T}};
+    kwargs...,
+) where {𝔽,O<:AbstractGroupOperation,T}
+    X = log(G.lie_group, e, T)
+    is_point(LieAlgebra(G), X; widthin=log, context=(:Output,), kwargs...)
+    return ValidationLieAlgebraTangentVector(X)
+end
 function ManifoldsBase.log!(G::ValidationLieGroup, X, g; kwargs...)
     is_point(G, g; widthin=log, context=(:Input,), kwargs...)
     log!(G.lie_group, unwrap_validation(X), unwrap_validation(g))