+++ test/operations/test_multiplication_operation.jl
@@ -52,7 +52,7 @@ using LieGroupsTestSuite
             LieGroupsTestSuite.DummyManifold(), AbelianMultiplicationGroupOperation()
         )
         # Edge case: Number element
-        @test identity_element(G, 0.0) == 1.0
+        @test identity_element(G, Float64) == 1.0
         # Edge case: Mixed compose
         @test compose(G, fill(1.0), 1.0) == 1.0
         @test compose(G, 1.0, fill(1.0)) == 1.0